netd: Prevent a failed exec netd from returning.
In order to start radish (IPv6 tethering) with the correct
permissions, netd forks and then execs the daemon. If the exec
step fails, (e.g. the binary not found, or no exec permissions
for radio user) the method was returning, which caused two instances
of netd to execute in parallel.

Exit instead of return.

Change-Id: I82ff7066657d76b22ab6a655a45a6e6a8cda8dce
CRs-Fixed: 295973