netd: Start listening to advertisements before sending solicitations
To retrieve the IPv6 gateway we send router solicitation packets. Then
we open a socket and wait for Router Advertisement packets.
It is possible that these packets are sent before we open out socket to
read them.

This change reduces the possibility of this race condition by opening
 and binding the listening socket before sending the router
solicitations.

Change-Id: I521eb6250868ca001110c193a279a34e085f5042
CRs-Fixed: 362789