netd: Support the Qualcomm SoftAP SDK
The Qualcomm SoftAP SDK extendes the HotSpot capability provided by
the stock Android distribution.  This allows the creation of a more
feature-rich HotSpot application.

In order to minimize the changes required to the stock Android
distribution, place the majority of the changes into a new
QualcommSoftapCmd class that contains most of the command dispatching
logic.

Change-Id: I3dd1690095f81820bf1d712f07198bba5b13629e