netd: Change QualcommSoftapCmd to add AR6004 support
Currently, commands for setting SoftAP will be redirected
to QualcommSoftapCmd, and this will set some fields in
the hostapd.conf that AR6004 driver doesn't support.

In this change, "set" command will be redirected to
the Android native SoftapController to let AR6004 able run
into softap mode, when the WLAN is AR6004.

Change-Id: I15951d99207a7ed67f2bf4986343e19bfdf3eb7f