netd: Support the QSAP Softap SDK.
The QSoftAP SDK extendes the HotSpot capability provided by
the stock Android distribution.  This allows the creation of a more
feature-rich HotSpot application.
In order to minimize the changes required to the stock Android
distribution, place the majority of the changes into a new
QsoftapCmd class that contains most of the command dispatching
logic.

Change-Id: Iaa895f11208641b72e2dc4ba7120f7dbef347577
CRs-Fixed: 436332