netd: Change QsoftapCmd to add AR6004 support
Currently, commands for setting SoftAP will be redirected
to QsoftapCmd, and this will set some fields in
the hostapd.conf that AR6004 driver doesn't support.
In this change, "set" command will be redirected to
the Android native SoftapController to let AR6004 able run
into softap mode, when the WLAN is AR6004.

Change-Id: Iec65332facf8ee84ef76d2b919eadbbec4d5cfdf
CRs-Fixed: 436332