netd: Start NDP proxy daemon only if atleast 2 interfaces are present
After repeatedly toggling tethering/Wifi hotspot on and off, hotspot may
not turn on anymore. It appears untethering is unsuccessful and this
eventually leads to a timeout and a framework reboot.
The NDP proxy daemon appears to be started multiple times even though
there is no ipv6 upstream interface present. There appears to be a
signifcant number of calls to fork and exec this daemon even though
this functionality is not needed in this case. Fix this overhead by
invoking the daemon only if there are atleast 2 interfaces to establish
a bridge and proxy the NDP messages.

CRs-fixed: 681922
Change-Id: I5ec3cdd82bcf90580e7db7dbacd3e0afe40cc044