netd: add support for ipv6 tethering when using dun apn
When tethering is enabled and a dun apn is used, a policy route
table is created per upstream interface. The table name is an
offset from BASE_TABLE_NUMBER. There is a rule added per upstream
interface which is specific for the tethered traffic which looks
up this table.
This patch adds the capability to pass this table name specific
to the upstream interface to the router advertisement daemon.
The route advertisement daemon will then add the rule which
looks up this table for the ipv6 traffic.

CRs-fixed: 654872
Change-Id: I2845558068ba53032fbbc6be618ac64cfed63c6e